Re-organized grid view tests.  Added more tests for more coverage.
